+++ src/XRoadLib/Serialization/CustomSerialization.cs
@@ -1,0 +1,10 @@
+﻿using System.Xml;
+
+namespace XRoadLib.Serialization
+{
+    public class CustomSerialization : ICustomSerialization
+    {
+        public virtual void OnContentComplete(XmlWriter writer)
+        { }
+    }
+}