+++ src/XRoadLib/Serialization/ICustomSerialization.cs
@@ -1,0 +1,9 @@
+﻿using System.Xml;
+
+namespace XRoadLib.Serialization
+{
+    public interface ICustomSerialization
+    {
+        void OnContentComplete(XmlWriter writer);
+    }
+}