+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+#### 1.0.0-alpha2 - Unreleased
+* Add SOAP specific helper methods.
+* Add helper methods to resolve operation contracts.
+
 #### 1.0.0-alpha1 - January 7 2016
 * Initial release
 * Extract common logic into separate project.